+++ app/src/main/java/my/edu/tarc/studicash_0703/IncomeHistoryActivity.kt
@@ -3,14 +3,16 @@ package my.edu.tarc.studicash_0703
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
-import androidx.activity.enableEdgeToEdge
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.ActivityResultCallback
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
-import my.edu.tarc.studicash_0703.Fragment.EditTransactionFragment
 import my.edu.tarc.studicash_0703.Models.Transaction
 import my.edu.tarc.studicash_0703.adapter.TransactionAdapter
 import my.edu.tarc.studicash_0703.databinding.ActivityIncomeHistoryBinding
@@ -21,15 +23,23 @@ class IncomeHistoryActivity : AppCompatActivity(), TransactionAdapter.OnTransact
     private val incomeTransactions = mutableListOf<Transaction>()
     private val db = FirebaseFirestore.getInstance()
 
+    private val addDurationActivityResultLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
+        if (result.resultCode == RESULT_OK) {
+            val data = result.data
+            val startDate = data?.getStringExtra("startDate")
+            val endDate = data?.getStringExtra("endDate")
+            binding.StartDate.text = startDate
+            binding.endDate.text = endDate
+            fetchTransactionsWithinDateRange(startDate, endDate)
+        }
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-
-        // Initialize binding
         binding = ActivityIncomeHistoryBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
+        ViewCompat.setOnApplyWindowInsetsListener(binding.main) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
@@ -42,9 +52,7 @@ class IncomeHistoryActivity : AppCompatActivity(), TransactionAdapter.OnTransact
         }
 
         binding.allButton.setOnClickListener {
-            val intent = Intent(this, TransactionHistoryActivity::class.java)
-            startActivity(intent)
-            finish()
+            fetchIncomeTransactions()
         }
 
         binding.expensesButton.setOnClickListener {
@@ -53,27 +61,29 @@ class IncomeHistoryActivity : AppCompatActivity(), TransactionAdapter.OnTransact
             finish()
         }
 
-        binding.incomesButtonMain.setOnClickListener {
-            fetchIncomeTransactions()
+        binding.AddDurationButton.setOnClickListener {
+            val intent = Intent(this, TransactionHistoryAddDurationActivity::class.java)
+            addDurationActivityResultLauncher.launch(intent)
         }
 
         binding.backBtn2.setOnClickListener {
             onBackPressed()
         }
+
         fetchIncomeTransactions()
     }
 
     private fun fetchIncomeTransactions() {
         val userId = FirebaseAuth.getInstance().currentUser?.uid ?: return
-        val expenseCollection = FirebaseFirestore.getInstance().collection("incomeTransactions")
+        val incomeCollection = FirebaseFirestore.getInstance().collection("incomeTransactions")
             .whereEqualTo("userId", userId)
 
-        expenseCollection.get()
+        incomeCollection.get()
             .addOnSuccessListener { result ->
                 incomeTransactions.clear()
                 for (document in result) {
                     val transaction = document.toObject(Transaction::class.java).copy(
-                        id = document.id // Ensure the ID is set from Firestore
+                        id = document.id
                     )
                     incomeTransactions.add(transaction)
                 }
@@ -85,38 +95,76 @@ class IncomeHistoryActivity : AppCompatActivity(), TransactionAdapter.OnTransact
             }
     }
 
+    private fun fetchTransactionsWithinDateRange(startDate: String?, endDate: String?) {
+        val userId = FirebaseAuth.getInstance().currentUser?.uid ?: return
+        val incomeCollection = FirebaseFirestore.getInstance().collection("incomeTransactions")
+            .whereEqualTo("userId", userId)
+            .whereGreaterThanOrEqualTo("date", startDate ?: "")
+            .whereLessThanOrEqualTo("date", endDate ?: "")
+
+        incomeCollection.get()
+            .addOnSuccessListener { result ->
+                incomeTransactions.clear()
+                for (document in result) {
+                    val transaction = document.toObject(Transaction::class.java).copy(
+                        id = document.id
+                    )
+                    incomeTransactions.add(transaction)
+                }
+                updateRecyclerView()
+            }
+            .addOnFailureListener { exception ->
+                Log.e(TAG, "Error fetching transactions within date range", exception)
+            }
+    }
+
     private fun updateRecyclerView() {
         val allTransactions = mutableListOf<Transaction>()
         allTransactions.addAll(incomeTransactions)
 
-        // Sort transactions by date in descending order
         allTransactions.sortByDescending { it.date }
 
-        // Update the adapter with the new dataset
         transactionAdapter.updateData(allTransactions)
     }
 
+    private fun showDeleteConfirmationDialog(transactionId: String, isExpense: Boolean) {
+        AlertDialog.Builder(this)
+            .setTitle("Delete Transaction")
+            .setMessage("Are you sure you want to delete this transaction?")
+            .setPositiveButton("Yes") { dialog, _ ->
+                deleteTransaction(transactionId, isExpense)
+                dialog.dismiss()
+            }
+            .setNegativeButton("No") { dialog, _ ->
+                dialog.dismiss()
+            }
+            .create()
+            .show()
+    }
+
+    override fun onDelete(transactionId: String, isExpense: Boolean) {
+        showDeleteConfirmationDialog(transactionId, isExpense)
+    }
+
+    override fun onEdit(transactionId: String) {
+        // Handle edit transaction
+    }
+
     private fun deleteTransaction(transactionId: String, isExpense: Boolean) {
-        val collection = FirebaseFirestore.getInstance().collection("incomeTransactions")
+        val collection = if (isExpense) {
+            FirebaseFirestore.getInstance().collection("expenseTransactions")
+        } else {
+            FirebaseFirestore.getInstance().collection("incomeTransactions")
+        }
 
         collection.document(transactionId).delete()
             .addOnSuccessListener {
-                fetchIncomeTransactions() // Refresh the list after deletion
+                Log.d(TAG, "Transaction deleted successfully.")
+                fetchIncomeTransactions() // Refresh incomes
             }
             .addOnFailureListener { exception ->
                 Log.e(TAG, "Error deleting transaction: ${exception.message}", exception)
             }
-    }
-
-    override fun onDelete(transactionId: String, isExpense: Boolean) {
-        deleteTransaction(transactionId, isExpense)
-    }
-
-    override fun onEdit(transactionId: String) {
-        // Handle the edit event, e.g., start an edit activity
-        val intent = Intent(this, EditTransactionFragment::class.java)
-        intent.putExtra("transactionId", transactionId)
-        startActivity(intent)
     }
 
     companion object {
